+++ tests/attrs.rs
@@ -1,106 +1,91 @@
-use hypo::{Attrs, Render};
+use hypo::{Attrs, render};
 
 #[test]
 fn single() {
-    let mut s = String::new();
-    Attrs(("id", "idê", ())).render(&mut s);
-    assert_eq!(s, " id=\"idê\">");
+    let s = Attrs(("id", "idê", ()));
+    assert_eq!(render(s).0, " id=\"idê\">");
 }
 
 #[test]
 fn many() {
-    let mut s = String::new();
-    Attrs(("value", 13, ("id", "idê", ()))).render(&mut s);
-    assert_eq!(s, " id=\"idê\" value=\"13\">");
+    let s = Attrs(("value", 13, ("id", "idê", ())));
+    assert_eq!(render(s).0, " id=\"idê\" value=\"13\">");
 }
 
 #[test]
 fn consecutive() {
-    let mut s = String::new();
-    Attrs(("class", "yellow", ("class", "red", ()))).render(&mut s);
-    assert_eq!(s, " class=\"red yellow\">");
+    let s = Attrs(("class", "yellow", ("class", "red", ())));
+    assert_eq!(render(s).0, " class=\"red yellow\">");
 }
 
 #[test]
 fn empty_single() {
-    let mut s = String::new();
-    Attrs(("value", (), ())).render(&mut s);
-    assert_eq!(s, ">");
+    let s = Attrs(("value", (), ()));
+    assert_eq!(render(s).0, ">");
 }
 
 #[test]
 fn empty_before() {
-    let mut s = String::new();
-    Attrs(("class", "red", ("class", (), ()))).render(&mut s);
-    assert_eq!(s, " class=\"red\">");
+    let s = Attrs(("class", "red", ("class", (), ())));
+    assert_eq!(render(s).0, " class=\"red\">");
 }
 
 #[test]
 fn empty_after() {
-    let mut s = String::new();
-    Attrs(("class", (), ("class", "red", ()))).render(&mut s);
-    assert_eq!(s, " class=\"red\">");
+    let s = Attrs(("class", (), ("class", "red", ())));
+    assert_eq!(render(s).0, " class=\"red\">");
 }
 
 #[test]
 fn empty_between() {
-    let mut s = String::new();
-    Attrs(("class", "yellow", ("class", (), ("class", "red", ())))).render(&mut s);
-    assert_eq!(s, " class=\"red yellow\">");
+    let s = Attrs(("class", "yellow", ("class", (), ("class", "red", ()))));
+    assert_eq!(render(s).0, " class=\"red yellow\">");
 }
 
 #[test]
 fn empty_between_skip() {
-    let mut s = String::new();
-    Attrs(("class", "yellow", ("id", (), ("class", "red", ())))).render(&mut s);
-    assert_eq!(s, " class=\"red yellow\">");
+    let s = Attrs(("class", "yellow", ("id", (), ("class", "red", ()))));
+    assert_eq!(render(s).0, " class=\"red yellow\">");
 }
 
 #[test]
 fn bool_true() {
-    let mut s = String::new();
-    Attrs(("checked", true, ())).render(&mut s);
-    assert_eq!(s, " checked>");
+    let s = Attrs(("checked", true, ()));
+    assert_eq!(render(s).0, " checked>");
 }
 
 #[test]
 fn bool_false() {
-    let mut s = String::new();
-    Attrs(("checked", false, ())).render(&mut s);
-    assert_eq!(s, ">");
+    let s = Attrs(("checked", false, ()));
+    assert_eq!(render(s).0, ">");
 }
 
 #[test]
 fn bool_true_true() {
-    let mut s = String::new();
-    Attrs(("checked", true, ("checked", true, ()))).render(&mut s);
-    assert_eq!(s, " checked>");
+    let s = Attrs(("checked", true, ("checked", true, ())));
+    assert_eq!(render(s).0, " checked>");
 }
 
 #[test]
 fn bool_true_false() {
-    let mut s = String::new();
-    Attrs(("checked", false, ("checked", true, ()))).render(&mut s);
-    assert_eq!(s, " checked>");
+    let s = Attrs(("checked", false, ("checked", true, ())));
+    assert_eq!(render(s).0, " checked>");
 }
 
 #[test]
 fn bool_false_true() {
-    let mut s = String::new();
-    Attrs(("checked", true, ("checked", false, ()))).render(&mut s);
-    assert_eq!(s, " checked>");
+    let s = Attrs(("checked", true, ("checked", false, ())));
+    assert_eq!(render(s).0, " checked>");
 }
 
 #[test]
 fn bool_false_false() {
-    let mut s = String::new();
-    Attrs(("checked", false, ("checked", false, ()))).render(&mut s);
-    assert_eq!(s, ">");
+    let s = Attrs(("checked", false, ("checked", false, ())));
+    assert_eq!(render(s).0, ">");
 }
 
 #[test]
 fn bool_true_false_true() {
-    let mut s = String::new();
-    Attrs(("checked", true, ("checked", false, ("checked", true, ())))).render(&mut s);
-    assert_eq!(s, " checked>");
+    let s = Attrs(("checked", true, ("checked", false, ("checked", true, ()))));
+    assert_eq!(render(s).0, " checked>");
 }